Added extension concatenation :wrench:
Remember, darlene only stores single file extensions, so if the input file was `/path/to/key.pub.sec`, the output file is `/path/to/key.drln`, and when decrypted it becomes `/path/to/key.sec` not `/path/to/key.pub.sec`.

To workaround this, the `-C` (or `--concat`) flag has been added. It allows the written file to maintain the file extension of the user provided output, then appends the extension stored by darlene.

So:

```sh
$ darlene -f /path/to/key.drln -o ~/Documents/key.pub -D -C
```

would write `~/Documents/key.pub.sec`.