utils: added darlene ver check
This is a best-effort attempt at determining whether an alledged drln file a legit one.

Note: There would likely be files that begin with a valid darlene ver but might not be a drln file, in that case, the user would just get additional errors about keylength and such.

We could add a darlene validity checker fn for both legacy and new drln files.

For the legacy checker it could just check that the ver, key length, encoding, IV, Tag, and ext are all in the expected value ranges or follow predictable patterns.

As for the new drln file format, we can just check the validity of the magic number and also perform a similar check to the legacy version's on the header data.